# doc/api.yaml
openapi: 3.0.0
info:
  title: Chat Application API
  description: Chat application REST API documentation
  version: 1.0.0
  contact:
    name: Chat App Team
    email: support@chatapp.com
    url: https://chatapp.com/support
servers:
- url: /api/v1
  description: API foundation path
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT authentication using the Bearer scheme.
        Example: "Authorization: Bearer {token}"
  responses:
    BadRequest:
      description: >
        The request was malformed or contained invalid parameters.
        Check the error details for specific validation issues.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 400
            message: "Invalid request parameters"
            trace: "req-123456"
            error: "validation.failed"
            errors:
              - "username: Must be between 3-50 characters"
              - "email: Invalid email format"
    Unauthorized:
      description: >
        Authentication failed or user lacks proper permissions.
        Include a valid JWT token in the Authorization header.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 401
            message: "Unauthorized: Invalid or expired token"
            trace: "req-789012"
            error: "auth.invalid_token"
            errors: []      
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 404
            message: "User not found"
            trace: "req-345678"
            error: "resource.not_found"
            errors: []     
    Forbidden:
      description: >
        Authenticated user does not have permission to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    User:
      type: object
      description: User information model containing all user-related data
      required:
      - id
      - username
      - avatar_url
      - email
      - gender
      properties:
        id:
          type: string
          description: Unique identifier for the user
          pattern: ^[a-zA-Z0-9-]+$
          minLength: 1
          maxLength: 64
        username:
          type: string
          description: User's display name
          minLength: 1
          maxLength: 50
          pattern: ^[a-zA-Z0-9_-]+$
        name:
          type: string
          description: User's full name
          minLength: 1
          maxLength: 100
          pattern: ^[a-zA-Z0-9\s'-]+$
        email:
          type: string
          description: User's email address
          format: email
          minLength: 5
          maxLength: 255
          pattern: ^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$
        avatar_url:
          type: string
          description: URL of user's avatar
          format: url
          maxLength: 255
          minLength: 1
          pattern: ^https?://[^\s/$.?#].[^\s]*$
        gender:
          type: string
          description: User's gender
          enum:
          - male
          - female
          - unspecified
          default: unspecified
...
    UserEnvelopeCollection:
      description: Envelope containing a list of users and optional paging info.
      allOf:
        - $ref: '#/components/schemas/BaseSuccessResponse'
        - type: object
          description: Success response with a list of user resources
          required:
            - items
          properties:
            items:
              type: array
              description: List of users
              minItems: 0
              maxItems: 100
              items:
                $ref: '#/components/schemas/User'
...
    FileUpload:
      type: object
      description: File upload information model
      required:
      - filename
      - url
      properties:
        filename:
          type: string
          description: Original filename
          minLength: 1
          maxLength: 255
          pattern: ^[a-zA-Z0-9_\-. ]+$
        size:
          type: integer
          format: int64
          description: File size in bytes
          minimum: 1
          maximum: 10485760  # 10MB
        url:
          type: string
          description: File access URL
          format: url
          minLength: 1
          maxLength: 255
          pattern: ^https?://[^\s/$.?#].[^\s]*$
    FileUploadEnvelope:
      description: Metadata and status returned after a file upload completes.
      allOf:
        - $ref: '#/components/schemas/BaseSuccessResponse'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              description: Payload containing uploaded file metadata
              required:
                - file
              properties:
                file:
                  $ref: '#/components/schemas/FileUpload'
...
    Error:
      type: object
      description: Standard error response model for API error conditions
      required:
      - code
      - message
      - trace
      - error
      - errors
      properties:
        code:
          type: integer
          format: int64
          description: HTTP error code
          minimum: 400
          maximum: 599
        message:
          type: string
          description: Human-readable error message
          minLength: 1
          maxLength: 255
          pattern: ^[\s\S]{1,255}$
        trace:
          type: string
          description: Error trace ID for debugging purposes
          pattern: ^[a-zA-Z0-9\s.,!?-]+$
          minLength: 1
          maxLength: 1000
        error:
          type: string
          description: Machine-readable error key (stable identifier for programmatic handling)
          minLength: 1
          maxLength: 100
          pattern: ^[a-z0-9_.:-]+$
          example: auth.invalid_token
        errors:
          type: array
          description: Collection of detailed error messages
          items:
            type: string
            minLength: 1
            maxLength: 255
            pattern: ^[\s\S]{1,255}$
          minItems: 0
          maxItems: 100
    BaseSuccessResponse:
      type: object
      description:  Base structure for all success responses
      required:
      - code
      - message
      properties:
        code:
          type: integer
          format: int64
          description: HTTP status code indicating success
          enum:
          - 200
          - 201
          minimum: 200
          maximum: 201
          example: 200
        message:
          type: string
          description: Human-readable success message
          minLength: 1
          maxLength: 255
          pattern: ^[\s\S]{1,1000}$
          example: "Operation succeeded"  
...
  parameters:
...
    id:
      name: id
      in: path
      required: true
      description: ID of the resource
      schema:
        type: string
        pattern: ^[a-zA-Z0-9-]+$
        minLength: 1
        maxLength: 64
        description: uniquie identifier
paths:
...
  /users/set_username:
    put:
      tags:
      - users
      operationId: setMyUserName
      summary: Change username
      description: Update the current user's username
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Request to update username
              properties:
                username:
                  type: string
                  description: New username
                  minLength: 1
                  maxLength: 50
                  pattern: ^[a-zA-Z0-9_-]+$
      responses:
        '200':
          description: Username updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseSuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/set_photo:
    put:
      tags:
      - users
      operationId: setMyPhoto
      summary: Change profile photo
      description: Update the current user's profile picture
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
              - upload
              description: Request to update profile photo
              properties:
                upload:
                  type: string
                  format: binary
                  description:  Image file (JPEG/PNG)
                  minLength: 1
                  maxLength: 10485760
      responses:
        '200':
          description: Profile photo updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileUploadEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /users/search:
    get:
      tags:
      - users
      summary: Search users by keyword
      description: Search users by partial name, username or email
      operationId: searchUsers
      parameters:
      - name: q
        in: query
        required: true
        description: Search query
        schema:
          type: string
          minLength: 1
          maxLength: 255
          pattern: ^[a-zA-Z0-9\s@.-]+$
          description: Search keyword to match against user names,usernames, or email addresses
      security:
      - BearerAuth: []
      responses:
        '200':
          description: Matching users found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEnvelopeCollection'
              example:
                code: 200
                message: Matching users found
                items:
                  - id: usr-1
                    username: alice
                    avatar_url: https://example.com/avatars/alice.jpg
                    email: alice@example.com
                    gender: female
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /groups/{id}/photo:
    parameters:
      - $ref: '#/components/parameters/id'
    put:
      tags:
      - groups
      operationId: setGroupPhoto
      summary: Change group photo
      description: Update the group's profile picture
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              description: Request to update group photo
              required:
                - upload
              properties:
                upload:
                  type: string
                  format: binary
                  description: Group photo file
                  minLength: 1
                  maxLength: 10485760
      responses:
        '200':
          description: Group photo updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileUploadEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...